requirements.txt
asgiref==3.3.1
boto3==1.16.56
botocore==1.19.56
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dataclasses==0.8
Django==3.1.5
django-filter==2.4.0
djangorestframework==3.12.2
filelock==3.0.12
freeze==3.0
idna==2.10
importlib-metadata==3.4.0
jmespath==0.10.0
joblib==1.0.0
Markdown==3.3.3
numpy==1.19.5
packaging==20.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-transformers==1.2.0
pytz==2020.5
regex==2020.11.13
requests==2.25.1
s3transfer==0.3.4
sacremoses==0.0.43
sentencepiece==0.1.95
six==1.15.0
sqlparse==0.4.1
tokenizers==0.9.4
torch==1.7.1
tqdm==4.56.0
transformers==4.2.1
typing-extensions==3.7.4.3
urllib3==1.26.2
zipp==3.4.0
